moving model-specific code to model.hpp
this makes main.cpp very small and simple to pre-generate
